process_monitor/requirements.txt
Django>=5.2.0
djangorestframework>=3.16.0
psutil>=7.0.0
requests>=2.32.0
pyinstaller>=6.15.0 